cdp/requirements.txt
asyncpg
sqlalchemy
google-cloud-sql-connector
colorama
python-dotenv
faker
